+++ app/src/main/java/protect/card_locker/preferences/SettingsActivity.java
@@ -17,27 +17,33 @@ import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatDelegate;
 import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.DialogFragment;
+import androidx.fragment.app.FragmentResultListener;
 import androidx.preference.ListPreference;
 import androidx.preference.Preference;
 import androidx.preference.PreferenceFragmentCompat;
+
+import kotlin.Suppress;
 import nl.invissvenska.numberpickerpreference.NumberDialogPreference;
 import nl.invissvenska.numberpickerpreference.NumberPickerPreferenceDialogFragment;
 import protect.card_locker.CatimaAppCompatActivity;
 import protect.card_locker.MainActivity;
 import protect.card_locker.R;
 import protect.card_locker.Utils;
+import protect.card_locker.databinding.SettingsActivityBinding;
 
 public class SettingsActivity extends CatimaAppCompatActivity {
 
+    private SettingsActivityBinding binding;
     private final static String RELOAD_MAIN_STATE = "mReloadMain";
     private SettingsFragment fragment;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        binding = SettingsActivityBinding.inflate(getLayoutInflater());
         setTitle(R.string.settings);
-        setContentView(R.layout.settings_activity);
-        Toolbar toolbar = findViewById(R.id.toolbar);
+        setContentView(binding.getRoot());
+        Toolbar toolbar = binding.toolbar;
         setSupportActionBar(toolbar);
         ActionBar actionBar = getSupportActionBar();
         if (actionBar != null) {
@@ -163,6 +169,11 @@ public class SettingsActivity extends CatimaAppCompatActivity {
         }
 
         @Override
+        @SuppressWarnings("deprecation") // setTargetFragment
+        // androidx.preference.PreferenceDialogFragmentCompat uses the deprecated method
+        // `getTargetFragment()`, which throws if `setTargetFragment()` isn't used before.
+        // While this isn't fixed on upstream, suppress the deprecation warning
+        // https://issuetracker.google.com/issues/181793702
         public void onDisplayPreferenceDialog(Preference preference) {
             if (preference instanceof NumberDialogPreference) {
                 NumberDialogPreference dialogPreference = (NumberDialogPreference) preference;