+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased - 115
 
 - Open image in gallery on long-press
+- Apply Material style to dialogs
+- Support creating card by sharing an image to Catima
+- Add quick spend button to card screen
 
 ## v2.20.0 - 114